Move some classes to separate files for future common use Move 	genRandomString to MyUtils 	JsonRequest, JsonResponse, JsonException to JSONRequestHandler.hpp
